% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{clean_tweets}
\alias{clean_tweets}
\title{Clean Tweets}
\usage{
clean_tweets(tweet_data, remove.mentions = TRUE, remove.hashtags = TRUE,
  remove.urls = TRUE, remove.retweets = TRUE, remove.numbers = FALSE,
  lowercase = TRUE)
}
\arguments{
\item{remove.mentions}{TRUE by default, controls whether to remove mentions. Can be set to FALSE to keep them.}

\item{remove.hashtags}{TRUE by default, controls whether to remove hashtags Can be set to FALSE to keep them.}

\item{remove.urls}{TRUE by default, controls whether to remove urls. Can be set to FALSE to keep them.}

\item{remove.retweets}{TRUE by default, controls whether to remove retweets. Can be set to FALSE to keep them.}

\item{remove.numbers}{FALSE by default, controls whether to remove numbers Can be set to TRUE to remove them.}

\item{lowercase}{TRUE by default, controls whether to convert all characters to lowercase. Can be set to FALSE to retain case.}

\item{tweets}{An input dataset of raw tweets, usually from search_tweets()}
}
\description{
This function takes a dataframe of raw tweets and performs some basic cleaning and tokenization. It returns 
the input data.frame, now with a new column for clean_text, the tweets after cleaning. It also returns
the emojis in the tweets in their own column, and a count of emojis used in each tweet, for convenience.
}

