% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{collect_follower_timelines}
\alias{collect_follower_timelines}
\title{Collects timelines of followers}
\usage{
collect_follower_timelines(users, nfollowers = 100, nstatus = 200,
  minstatus = 20, language = "en")
}
\arguments{
\item{users}{The list of users whose followers you wish to collect.}

\item{nfollowers}{Defaults to 100. The maximum number of followers you wish to collect for each user.}

\item{nstatus}{Defaults to 200. The maximum number of statuses you wish to collect for each follower.}

\item{minstatus}{Defaults to 20. The minimum number of tweets a follower must have for you to download their timeline.}

\item{language}{Defaults to 'en'. A filter to only collect users of the specified language.}
}
\description{
This function takes a vector of users (e.g. c('seanchrismurphy', LydiaHayward2', 'katiehgreenaway)), and then collects the timelines of their 
followers, returning a dataset with up to nstatus tweets for each of nfollowers, for ever user entered. For instance if you input five
users with nfollowers = 100 and ntatus = 200, you would retrieve up to 100, 000 tweets (though usually substantially less, do to sparse
timelines)
}

