% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dictionary_count}
\alias{dictionary_count}
\title{Dictionary count}
\usage{
dictionary_count(tweets, group = "tweet")
}
\arguments{
\item{tweets}{This is the input data.}

\item{group}{Defaults to 'tweet', can be set to 'group' when working with timelines. If set to 'tweet' when working with timelines, will return one row per tweet.}
}
\description{
This function takes a data frame of tweets that has been cleaned with clean_tweets, and compares each word to the NRC lexicon.
Depending on the group parameter, it will then un-tokenize the tweets, giving either a clean tweet per row (if group is set
# to "tweet") or a clean user timeline per row (if group is set to "user").
}

