% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_Followers}
\alias{get_Followers}
\title{Get Followers}
\usage{
get_Followers(user)
}
\description{
This is a convenience function to get user followers a little more cleanly
than the default in the twitteR package.
}

