% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_Friends}
\alias{get_Friends}
\title{Get Friends}
\usage{
get_Friends(user)
}
\description{
This is a convenience function to get user friends a little more cleanly
than the default in the twitteR package.
}

