% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_timelines}
\alias{get_timelines}
\title{Get timelines}
\usage{
get_timelines(users, nstatus = 200, includeRts = FALSE)
}
\arguments{
\item{nstatus}{Defaults to 200. This is the maximum number of tweets that will be returned for each user.}

\item{includeRts}{Defaults to FALSE. This determines whether tweets that the user has retweeted will be returned. These tweets
are included in nstatus no matter what (this is a limitation of the Twitter API), so when includeRts is set to FALSE, some number less
than nstatus tweets will be returned for each user, depending on how many retweets they have made.}

\item{tweets}{This is the input data.}
}
\description{
This function takes a vector of users (e.g. of the form c('seanchrismurphy', 'lydiahayward2', 'katiegreenaway')) and returns
a data frame containing up to 3200 tweets of each of them, depending on nstatus.
}

