% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{most.frequent.mentions}
\alias{most.frequent.mentions}
\title{See most frequent mentions in the data}
\usage{
most.frequent.mentions(text)
}
\arguments{
\item{text}{is the input data.}
}
\description{
This function takes a tweet dataset and returns a data.frame showing the most frequent mentions in descending
order. Useful when checking to see if current events or spam are disturbing the data.
}

