% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{most.frequent.words}
\alias{most.frequent.words}
\title{See most frequent words in the data}
\usage{
most.frequent.words(text)
}
\arguments{
\item{text}{is the input data.}
}
\description{
This function takes a tweet dataset and returns a data.frame showing the most frequent word tokens (i.e.
# not hashtags or mentions) in descending order. Useful when checking to see if current events or spam are 
disturbing the data.
}

