% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tokenize_tweets}
\alias{tokenize_tweets}
\title{Utility function to tokenize tweets without removing anything. Used by the most.frequent.x functions.}
\usage{
tokenize_tweets(tweet_data, remove.words = FALSE, remove.hashtags = FALSE,
  remove.mentions = FALSE)
}
\description{
Utility function to tokenize tweets without removing anything. Used by the most.frequent.x functions.
}

