% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tot_limits}
\alias{tot_limits}
\title{Get total rate limits}
\usage{
tot_limits(query = NULL, token = get_tokens())
}
\arguments{
\item{query}{Defaults to NULL, which retrieves all limits, but can be used to specify a query such as 
'application/ratestatus' to only get a specific set of limits.}

\item{token}{This allows you to specify which token to check on, if you only want to check on one. Default is
to check all}
}
\description{
This utility function gets the requests remaining across multiple Twitter authentication tokens. Likely to 
be made redundant by updated to rtweet in the near future.
}

