% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{twtools-package}
\alias{twtools}
\alias{twtools-package}
\title{Utility functions for dealing with tweet data from twitteR.}
\description{
twtools is built to accompany the paper "A hands-on guide to accessing Twitter for psychological research"
}
\details{
The key functions you're likely to need from \pkg{twtools} are
\code{\link{clean_tweets}}, \code{\link{dictionary_count}}, and
\code{\link{create_closed_network}}
}

