% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create_closed_network}
\alias{create_closed_network}
\title{Create Closed Network}
\usage{
create_closed_network(usernames, max.followers = 5000,
  retryOnRateLimit = 120)
}
\arguments{
\item{usernames}{A vector of usernames.}

\item{max.followers}{Defaults to 5000, anyone with more followers than this will be ignored.}

\item{retryOnRateLimit}{is set to 120 by default, meaning the function will run until completion.}
}
\description{
This function takes a vector of users (e.g. of the form c('seanchrismurphy', 'lydiahayward2', 'katiegreenaway')) and returns
a data frame of followership links between those users. By default, it will ignore users with more than 5000 followers, to save
on computing time and remove celebrities. However this behavior can be changed with max.followers.
}

