% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_Followers}
\alias{get_Followers}
\title{Get Followers}
\usage{
get_Followers(users)
}
\description{
This is a convenience function to get user followers a little more cleanly
than the default in the twitteR package. It takes a list of user objects
returned by lookupUsers, and returns a dataset of user information for the
followers of those users. It attached a 'following' variable to the dataset
to track who is following who.
}

